[Math] drag solver with local memory
